+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2018-09-04
 
+- hmm.classify: show progress with colored dots depending on ranked models
 - adjustments/clarifications related with codebook's `raas`
 
 2018-09-03